+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram/fhi-diagram-options.model.ts
@@ -1,4 +1,4 @@
-import { DiagramType } from "./diagram-type.model";
+import { DiagramType } from "./fhi-diagram-type.model";
 
 interface Data {
   name: string;
