+++ projects/fhi-angular-highcharts/src/lib/fhi-diagram/fhi-diagram-type-id.ts
